# src/AuctionApp.Presentation/Controllers/UserWatchlistsController.cs
using Application.App.Queries;
using Application.App.UserWatchlists.Commands;
using Application.App.UserWatchlists.Responses;
using AuctionApp.Presentation.Common.Requests;
using MediatR;
using Microsoft.AspNetCore.Authorization;
using Microsoft.AspNetCore.Mvc;
using Presentation.Common.Abstractions;
using Swashbuckle.AspNetCore.Annotations;

namespace Presentation.Controllers;

[ApiController]
[Route("[controller]")]
[Authorize]
public class UserWatchlistsController : AppBaseController
{
    private readonly IMediator _mediator;

    public UserWatchlistsController(IMediator mediator)
    {
        _mediator = mediator;
    }

    [HttpGet("{id}")]
    [SwaggerOperation(OperationId = nameof(GetUserWatchlist))]
    public async Task<ActionResult<UserWatchlistDto>> GetUserWatchlist(int id)
    {
        var userId = GetUserId();

        var userWatchlistDto = await _mediator.Send(new GetUserWatchlistByIdQuery() { Id = id, UserId = userId });

        return Ok(userWatchlistDto);
    }

    [HttpGet]
    [SwaggerOperation(OperationId = nameof(GetUserWatchlistByProductId))]
    public async Task<ActionResult<UserWatchlistDto>> GetUserWatchlistByProductId([FromQuery] int productId)
    {
        var userId = GetUserId();

        var userWatchlistDto = await _mediator.Send(new GetUserwatchlistByProductIdQuery() { ProductId = productId, UserId = userId });

        return Ok(userWatchlistDto);
    }

    [HttpPost]
    [SwaggerOperation(OperationId = nameof(CreateUserWatchlist))]
    public async Task<ActionResult<UserWatchlistDto>> CreateUserWatchlist(CreateUserWatchlistRequest createUserWatchlistRequest)
    {
        var userId = GetUserId();

        var command = new CreateUserWatchlistCommand
        {
            UserId = userId,
            ProductId = createUserWatchlistRequest.ProductId,
        };

        var userWatchlistDto = await _mediator.Send(command);

        return Ok(userWatchlistDto);
    }

    [HttpDelete("{id}")]
    [SwaggerOperation(OperationId = nameof(DeleteUserWatchlist))]
    public async Task<ActionResult> DeleteUserWatchlist(int id)
    {
        var userId = GetUserId();

        await _mediator.Send(new DeleteUserWatchlistCommand() { Id = id, UserId = userId });

        return Ok();
    }
}